implementation of various authorizations in the controller allowing authorized or not the removal of a task

+++ src/Controller/DeleteTaskController.php
@@ -8,6 +8,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Security;
 
 class DeleteTaskController
 {
@@ -17,15 +18,22 @@ class DeleteTaskController
      * @param UrlGeneratorInterface $generator
      * @param Session $session
      * @param EntityManagerInterface $em
+     * @param Security $security
      * @return RedirectResponse
      */
     public function deleteTaskAction(Task $task, UrlGeneratorInterface $generator, Session $session,
-                                     EntityManagerInterface  $em)
+                                     EntityManagerInterface  $em, Security $security)
     {
-        $em->remove($task);
-        $em->flush();
+        if (($security->getUser() == $task->getUser()) ||
+            ($security->isGranted('ROLE_ADMIN') && $task->getUser() == null)) {
 
-        $session->getFlashBag()->add('success', 'La tâche a bien été supprimée.');
+            $em->remove($task);
+            $em->flush();
+            $session->getFlashBag()->add('success', 'La tâche a bien été supprimée.');
+
+        } else {
+            $session->getFlashBag()->add('error', 'Désolé mais vous ne pouvez pas supprimer la tache.');
+        }
 
         $router = $generator->generate('task_list');
